Some work on Proline-Core: * synchronization with the latest mzdb-processing API * fix issue related to the loading of the ScanSequence (use the PK instead of the raw file name)
Former-commit-id: a1deb784b1bf8cc37a65f07ad77f02f1baa33fbd

+++ Proline-DAL/src/main/scala/fr/proline/core/om/provider/lcms/impl/AbstractSQLLcMsMapProvider.scala
@@ -141,7 +141,7 @@ abstract class AbstractSQLLcMsMapProvider extends ILcMsMapProvider {
        charge = ftRecord.getInt(FtCols.CHARGE),
        elutionTime = toFloat(ftRecord.getAny(FtCols.ELUTION_TIME)),
        duration = toFloat(ftRecord.getAny(FtCols.DURATION)),
-       qualityScore = ftRecord.getDoubleOrElse(FtCols.QUALITY_SCORE,Double.NaN),
+       qualityScore = ftRecord.getDoubleOrElse(FtCols.QUALITY_SCORE,Double.NaN).toFloat,
        ms1Count = ftRecord.getInt(FtCols.MS1_COUNT),
        ms2Count = ftRecord.getInt(FtCols.MS2_COUNT),
        isOverlapping = toBoolean(ftRecord.getAny(FtCols.IS_OVERLAPPING)),
