+++ Proline-OMP/src/main/scala/fr/proline/core/parser/lcms/impl/MFPaQMapParser.scala
@@ -62,7 +62,7 @@ class MFPaQMapParser extends ILcmsMapFileParser {
           elutionTime = apexRt,
           duration = 0, // FIXME
           charge = charge,
-          qualityScore = Double.NaN,
+          qualityScore = Float.NaN,
           ms1Count = lastScan.cycle - firstScan.cycle + 1,
           ms2Count = ms2EventIds.length,
           isOverlapping = false,
